Move default callbacks to scheduler mod
Moves the default result and error callbacks from the serial scheduler
to the top-level scheduler module so they can be used by other
schedulers.
